run command execs into docker-compose
The use of exec gets rid of the bash process and allows signals to be propagated
into the running container, instead of being dropped by bash.

This is important to allow graceful handling of timeout, wherein a TERM signal
is sent by BK to give the running job the opportunity to tidy up and exit
cleanly before a KILL signal is ultimately sent.